requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
plotly>=5.20
google-genai>=0.3
pydantic>=2.0
openpyxl>=3.1
kaleido>=0.2.1
imageio>=2.34
imageio-ffmpeg>=0.5
